Split tests into separate files and add more tests, still to fix
