+++ lib/openHAB/ruleBasedInterpreter.js
@@ -230,7 +230,7 @@ function itemProperties(expression, tags, groupRequired, itemType) {return item(
  * @param {boolean} [options.include=true]
  *     Whether to include matched items in execution parameters.
  * @param {boolean} [options.matchMultiple=false]
- *     Whether to allow multiples item matches (default: false = match only one item).
+ *     Whether to allow multiple item matches (default: false = match only one item).
  * @param {string|string[]} [options.type=null]
  *     Restrict to specific item type(s). Single string = exact match, array = match any.
  * @param {string|string[]} [options.tag=[]]
@@ -346,8 +346,8 @@ class ItemExp extends Expression {
                     return new EvaluationResult(false, tokens, null, null);
                 }
 
-                    if (this.groupContext === "one" && groups.length > 1) {
-                        logger.warn(`eval item fail: groupContext 'one' but multiple groups matched (${groups.length}); groups=[${groups.map(g=>g.name||g.label||'<unknown>').join(', ')}]`);
+                if (this.groupContext === "one" && groups.length > 1) {
+                    logger.warn(`eval item fail: groupContext 'one' but multiple groups matched (${groups.length}); groups=[${groups.map(g => g.name).join(", ")}]`);
                     return new EvaluationResult(false, tokens, null, null);
                 }
 
@@ -760,21 +760,30 @@ function getSynonyms(item) {
 }
 
 function itemIsInSubGroup(item, targetGroupItem) {
-    let checkedGroupNames = [];
-    let groupStack = [];
-    groupStack.push(item);
+    // Don't match the target group itself
+    if (item.name === targetGroupItem.name) {
+        return false;
+    }
+
+    let checkedGroupNames = new Set();
+    let groupStack = [...(item.groupNames || [])];
+
     while (groupStack.length > 0) {
-        let groupItem = groupStack.pop();
-        if (groupItem.name == targetGroupItem.name) {
+        const groupName = groupStack.pop();
+        
+        // Direct parent group matches
+        if (groupName === targetGroupItem.name) {
             return true;
         }
 
-        groupItem.groupNames.forEach(groupName => {
-            if (!checkedGroupNames.includes(groupName)) {
-                checkedGroupNames.push(groupName);
-                groupStack.push(items.getItem(groupName));
+        // Add ancestor groups if not already checked
+        if (!checkedGroupNames.has(groupName)) {
+            checkedGroupNames.add(groupName);
+            const group = items.getItem(groupName);
+            if (group && Array.isArray(group.groupNames)) {
+                groupStack.push(...group.groupNames);
             }
-        });
+        }
     }
 
     return false;
