+++ test/rbi.seq.test.js
@@ -1,0 +1,124 @@
+const {when} = require("jest-when");
+jest.mock("openhab");
+const openhab = require("openhab");
+openhab.items = {
+    getItems: jest.fn().mockReturnValue([]),
+    getItemsByTag: jest.fn().mockReturnValue([]),
+    getItem: jest.fn().mockReturnValue(null)
+};
+openhab.log = jest.fn().mockReturnValue({
+    /* Uncomment to get logs during test execution
+    info: console.log,
+    debug: console.debug,
+    warn: console.warn
+    //*/
+    ///* Comment to get logs during test execution
+    info: jest.fn(),
+    debug: jest.fn(),
+    warn: jest.fn()
+    //*/
+});
+const { RuleBasedInterpreter, seq, cmd, item } = require("../lib/openHAB/ruleBasedInterpreter");
+
+let rbi = new RuleBasedInterpreter();
+
+beforeEach(() => {
+    rbi.clearRules();
+    openhab.items.getItems.mockReturnValue([]);
+});
+
+describe("single sequence expression", () => {
+    it("matches correct sequence", () => {
+        let testExpression = seq("bar", "foo", "foobar");
+        let testFunction = jest.fn();
+        rbi.addRule(testExpression, testFunction);
+        rbi.interpretUtterance("bar foo foobar");
+        expect(testFunction.mock.calls.length).toBe(1);
+    });
+
+    it("does not match wrong sequence", () => {
+        let testExpression = seq("bar", "foo", "foobar");
+        let testFunction = jest.fn();
+        rbi.addRule(testExpression, testFunction);
+        rbi.interpretUtterance("foo foobar");
+        expect(testFunction.mock.calls.length).toBe(0);
+    });
+
+    it("matches normalized unicode characters and apostrophes", () => {
+        let testExpression = seq("tÈS$Tö'ó");
+        let testFunction = jest.fn();
+        rbi.addRule(testExpression, testFunction);
+        rbi.interpretUtterance("2TèstÖ'Ó");
+        expect(testFunction.mock.calls.length).toBe(1);
+    });
+
+    it("matches first token and ignores further", () => {
+        let testExpression = seq("foo bar");
+        let testFunction = jest.fn();
+        rbi.addRule(testExpression, testFunction);
+        rbi.interpretUtterance("foo abc");
+        expect(testFunction.mock.calls.length).toBe(1);
+    });
+
+    it("merges executeParameter from sequential expressions", () => {
+        const itemA = {name: "itemA", label: "itemA", sendCommand: jest.fn()};
+        const itemB = {name: "itemB", label: "itemB", sendCommand: jest.fn()};
+
+        // Mock items to be returned by different expressions
+        when(openhab.items.getItems).mockReturnValue([itemA, itemB]);
+
+        const testExpression = seq(item(), item(), cmd("foo", 123));
+        rbi.addRule(testExpression);
+
+        // Execute and verify both items received commands
+        rbi.interpretUtterance("itemA itemB foo");
+        expect(itemA.sendCommand).toHaveBeenCalledWith(123);
+        expect(itemB.sendCommand).toHaveBeenCalledWith(123);
+    });
+
+    it("preserves duplicates when merging executeParameter.items", () => {
+        const itemA = {name: "itemA", label: "itemA", sendCommand: jest.fn()};
+        when(openhab.items.getItems).mockReturnValue([itemA]);
+
+        const testExpression = seq(item(), item(), cmd("foo", 123));
+        rbi.addRule(testExpression);
+
+        // Item should receive command twice since it's included twice
+        rbi.interpretUtterance("itemA itemA foo");
+        expect(itemA.sendCommand).toHaveBeenCalledTimes(2);
+    });
+
+    describe("executeFunction precedence", () => {
+        it("selects the first non-null executeFunction from child expressions", () => {
+            const item1 = {name: "item1", label: "item1", sendCommand: jest.fn()};
+            const item2 = {name: "item2", label: "item2", sendCommand: jest.fn()};
+            
+            when(openhab.items.getItems).mockReturnValue([item1, item2]);
+
+            const func1 = jest.fn();
+            const func2 = jest.fn();
+
+            // First expression provides executeFunction
+            const expr1 = item({ label: "item1" });
+            const expr2 = item({ label: "item2" });
+            
+            rbi.addRule(seq(expr1, expr2), func1);
+            rbi.interpretUtterance("item1 item2");
+            
+            expect(func1).toHaveBeenCalled();
+            expect(func2).not.toHaveBeenCalled();
+
+            // Reset
+            func1.mockClear();
+            func2.mockClear();
+            rbi.clearRules();
+
+            // Second expression provides executeFunction
+            rbi.addRule(seq(expr2, expr1), func2);
+            rbi.interpretUtterance("item2 item1");
+            
+            expect(func2).toHaveBeenCalled();
+            expect(func1).not.toHaveBeenCalled();
+        });
+    });
+});
