--- test/ruleBasedInterpreter.test.js
@@ -1,727 +1,0 @@
-const {when} = require("jest-when");
-jest.mock("openhab");
-const openhab = require("openhab");
-openhab.items = {
-    getItems: jest.fn().mockReturnValue([]),
-    getItemsByTag: jest.fn().mockReturnValue([]),
-    getItem: jest.fn().mockReturnValue(null)
-};
-openhab.log = jest.fn().mockReturnValue({
-    /* Uncomment to get logs during test execution
-    info: console.log,
-    debug: console.debug,
-    warn: console.warn
-    //*/
-    ///* Comment to get logs during test execution
-    info: jest.fn(),
-    debug: jest.fn(),
-    warn: jest.fn()
-    //*/
-});
-const { RuleBasedInterpreter, alt, seq, opt, cmd, itemLabel, itemProperties } = require("../lib/openHAB/ruleBasedInterpreter");
-
-let rbi = new RuleBasedInterpreter();
-
-beforeEach(() => {
-    rbi.clearRules();
-    openhab.items.getItems.mockReturnValue([]);
-});
-
-describe("interpretUtterance", () => {
-    describe("single alternative expression", () => {
-        it("matches same string", () => {
-            let testExpression = alt("bar", "foo", "foobar");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("foo");
-            expect(testFunction.mock.calls.length).toBe(1);
-            rbi.interpretUtterance("bar");
-            expect(testFunction.mock.calls.length).toBe(2);
-            rbi.interpretUtterance("foobar");
-            expect(testFunction.mock.calls.length).toBe(3);
-        });
-
-        it("does not match other string", () => {
-            let testExpression = alt("bar", "foo", "foobar");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("fob");
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-    });
-
-    describe("command expression", () => {
-        it("executes no function when no item expression is contained", () => {
-            let testExpression = cmd("foo", 1);
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found item", () => {
-            let retItem = {label: "my item", sendCommand: jest.fn()};
-            openhab.items.getItems.mockReturnValue([retItem])
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(),cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("my item foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(retItem.sendCommand.mock.calls.length).toBe(1);
-            expect(retItem.sendCommand.mock.calls[0].length).toBe(1);
-            expect(retItem.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-        });
-    });
-
-    describe("item properties expression", () => {
-        it("sends command to found items by single tag", () => {
-            let itemOne = {sendCommand: jest.fn()};
-            let itemTwo = {sendCommand: jest.fn()};
-            let itemThree = {sendCommand: jest.fn()};
-            let itemFour = {sendCommand: jest.fn()};
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree, itemFour]);
-            when(openhab.items.getItemsByTag).calledWith(...["bar"]).mockReturnValue([itemOne, itemThree]);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties("something", ["bar"], false),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemFour.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found items by multiple tags", () => {
-            let itemOne = {sendCommand: jest.fn()};
-            let itemTwo = {sendCommand: jest.fn()};
-            let itemThree = {sendCommand: jest.fn()};
-            let itemFour = {sendCommand: jest.fn()};
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree, itemFour]);
-            when(openhab.items.getItemsByTag).calledWith(...["foo", "bar"]).mockReturnValue([itemOne, itemThree]);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties("something", ["foo", "bar"], false),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemFour.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found items by type", () => {
-            let itemOne = {type: "foobar", sendCommand: jest.fn()};
-            let itemTwo = {type: "foo", sendCommand: jest.fn()};
-            let itemThree = {type: "bar", sendCommand: jest.fn()};
-            let itemFour = {type: "foo", sendCommand: jest.fn()};
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree, itemFour]);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties("something", null, false, "foo"),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(0);
-            expect(itemFour.sendCommand.mock.calls.length).toBe(1);
-            expect(itemFour.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemFour.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found items by type and tags", () => {
-            let itemOne = {type: "foobar", sendCommand: jest.fn()};
-            let itemTwo = {type: "foo", sendCommand: jest.fn()};
-            let itemThree = {type: "bar", sendCommand: jest.fn()};
-            let itemFour = {type: "foo", sendCommand: jest.fn()};
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree, itemFour]);
-            when(openhab.items.getItemsByTag).calledWith(...["foo", "bar"]).mockReturnValue([itemOne, itemTwo]);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties("something", ["foo", "bar"], false, "foo"),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(0);
-            expect(itemFour.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found items by direct location and tag", () => {
-            let itemOne = {name: "itemOne", label: "itemOne", groupNames: ["itemTwo"], sendCommand: jest.fn()};
-            let itemTwo = {name: "itemTwo", label: "itemTwo", type: "Group", groupNames: ["itemThree"], sendCommand: jest.fn()};
-            let itemThree = {name: "itemThree", label: "itemThree", type: "Group", groupNames: [], sendCommand: jest.fn()};
-            
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-            when(openhab.items.getItemsByTag).calledWith(...["foo"]).mockReturnValue([itemOne, itemThree]);
-            when(openhab.items.getItem).calledWith("itemTwo").mockReturnValue(itemTwo);
-            when(openhab.items.getItem).calledWith("itemThree").mockReturnValue(itemThree);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties(seq("something",itemLabel(false, true)), ["foo"], true),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something itemTwo works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found items by sub location and tag", () => {
-            let itemOne = {name: "itemOne", label: "itemOne", groupNames: ["itemTwo"], sendCommand: jest.fn()};
-            let itemTwo = {name: "itemTwo", label: "itemTwo", type: "Group", groupNames: ["itemThree"], sendCommand: jest.fn()};
-            let itemThree = {name: "itemThree", label: "itemThree", type: "Group", groupNames: [], sendCommand: jest.fn()};
-            
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-            when(openhab.items.getItemsByTag).calledWith(...["foo"]).mockReturnValue([itemOne]);
-            when(openhab.items.getItem).calledWith("itemTwo").mockReturnValue(itemTwo);
-            when(openhab.items.getItem).calledWith("itemThree").mockReturnValue(itemThree);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties(seq("something",itemLabel(false, true)), ["foo"], true),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something itemThree works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found items by sub location and tag and type", () => {
-            let itemOne = {name: "itemOne", label: "itemOne", type: "foobar", groupNames: ["itemTwo"], sendCommand: jest.fn()};
-            let itemTwo = {name: "itemTwo", label: "itemTwo", type: "Group", groupNames: ["itemThree"], sendCommand: jest.fn()};
-            let itemThree = {name: "itemThree", label: "itemThree", type: "Group", groupNames: [], sendCommand: jest.fn()};
-            
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-            when(openhab.items.getItemsByTag).calledWith(...["foo"]).mockReturnValue([itemOne]);
-            when(openhab.items.getItem).calledWith("itemTwo").mockReturnValue(itemTwo);
-            when(openhab.items.getItem).calledWith("itemThree").mockReturnValue(itemThree);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties(seq("something",itemLabel(false, true)), ["foo"], true, "foobar"),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something itemThree works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends command to found items by self location and tag", () => {
-            let itemOne = {name: "itemOne", label: "itemOne", type: "Group", groupNames: ["itemTwo"], sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let itemTwo = {name: "itemTwo", label: "itemTwo", type: "Group", groupNames: ["itemThree"], sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let itemThree = {name: "itemThree", label: "itemThree", type: "Group", groupNames: [], sendCommand: jest.fn(), getMetadata: jest.fn()};
-            
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-            when(openhab.items.getItemsByTag).calledWith(...["foo"]).mockReturnValue([itemOne, itemThree]);
-            when(openhab.items.getItem).calledWith("itemTwo").mockReturnValue(itemTwo);
-            when(openhab.items.getItem).calledWith("itemThree").mockReturnValue(itemThree);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties(seq("something",itemLabel(false, true)), ["foo"], true),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something itemOne works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("sends no command if items are not in location", () => {
-            let itemOne = {name: "itemOne", label: "itemOne", type: "Switch", groupNames: ["itemTwo"], sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let itemTwo = {name: "itemTwo", label: "itemTwo", type: "Group", groupNames: ["itemThree"], sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let itemThree = {name: "itemThree", label: "itemThree", type: "Group", groupNames: [], sendCommand: jest.fn(), getMetadata: jest.fn()};
-            
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-            when(openhab.items.getItemsByTag).calledWith(...["foo"]).mockReturnValue([itemOne, itemThree]);
-            when(openhab.items.getItem).calledWith("itemTwo").mockReturnValue(itemTwo);
-            when(openhab.items.getItem).calledWith("itemThree").mockReturnValue(itemThree);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemProperties(seq("something",itemLabel(false, true)), ["foo"], true),cmd("works", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("something itemOne works");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(0);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("turns lights on in location", () => {
-            let itemOne = {name: "itemOne", label: "itemOne", type: "Switch", groupNames: ["itemTwo"], sendCommand: jest.fn()};
-            let itemTwo = {name: "itemTwo", label: "itemTwo", type: "Group", groupNames: ["itemThree"], sendCommand: jest.fn()};
-            let itemThree = {name: "itemThree", label: "itemThree", type: "Group", groupNames: [], sendCommand: jest.fn()};
-            
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-            when(openhab.items.getItemsByTag).calledWith(...["Light"]).mockReturnValue([itemOne]);
-            when(openhab.items.getItem).calledWith("itemTwo").mockReturnValue(itemTwo);
-            when(openhab.items.getItem).calledWith("itemThree").mockReturnValue(itemThree);
-            
-            let onOff = alt(cmd("on", "ON"), cmd("off", "OFF"));
-            let turn = alt("turn", "switch");
-            let the = opt("the");
-            let inOfThe = seq(alt("in", "of"), the);
-            let allThe = alt(seq("all", the), the);
-            let lights = alt("light", "lights");
-            let testExpression = seq(
-                turn,
-                opt(onOff),
-                allThe,
-                itemProperties(
-                    seq(
-                        lights,
-                        opt(onOff),
-                        inOfThe,
-                        itemLabel(false, true)),
-                    ["Light"], true, "Switch"),
-                opt(onOff))
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("turn on all the lights in itemThree");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe("ON");
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(0);
-        });
-    });
-
-    describe("item expression", () => {
-        it("sends command to found item by label", () => {
-            let itemOne = {label: "item one", sendCommand: jest.fn()};
-            let itemTwo = {label: "item two", sendCommand: jest.fn()};
-            let itemThree = {label: "third item", sendCommand: jest.fn()};
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-            
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(),cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("item one foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-
-            rbi.interpretUtterance("item two foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-
-            rbi.interpretUtterance("third item foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-        });
-
-        it("sends command to found item by synonym", () => {
-            let firstName = "item one";
-            let secondName = "item two";
-            let thirdName = "third item";
-            let getMetadataImplementation = (data) => (namespace) => {
-                if (namespace != "synonyms") {
-                    return null;
-                }
-                return {value: data};
-            };
-
-            let itemOne = {label: firstName, sendCommand: jest.fn(),
-                getMetadata: jest.fn(getMetadataImplementation("bar one, barbar"))};
-            let itemTwo = {label: secondName, sendCommand: jest.fn(),
-                getMetadata: jest.fn(getMetadataImplementation("second bar"))};
-            let itemThree = {label: thirdName, sendCommand: jest.fn(),
-                getMetadata: jest.fn(getMetadataImplementation("bar three"))};
-            openhab.items.getItems.mockReturnValue([itemOne, itemTwo, itemThree]);
-
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(),cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            
-            rbi.interpretUtterance("bar one foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-
-            rbi.interpretUtterance("barbar foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemOne.sendCommand.mock.calls.length).toBe(2);
-            expect(itemOne.sendCommand.mock.calls[1].length).toBe(1);
-            expect(itemOne.sendCommand.mock.calls[1][0]).toBe(cmdParameter);
-
-            rbi.interpretUtterance("second bar foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemTwo.sendCommand.mock.calls.length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemTwo.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-
-            rbi.interpretUtterance("bar three foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            expect(itemThree.sendCommand.mock.calls.length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0].length).toBe(1);
-            expect(itemThree.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-        });
-
-        it("prefers the shortest exact matching label when both short and long labels match the utterance", () => {
-            let shortItem = {label: "TV", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let longItem = {label: "TV Room", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            openhab.items.getItems.mockReturnValue([shortItem, longItem]);
-
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(), opt("Room"), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            // Both items match the utterance "TV Room foo" (short label "TV" and long label "TV Room").
-            // Shortest exact match ("TV") should be preferred.
-            rbi.interpretUtterance("TV Room foo");
-            expect(shortItem.sendCommand.mock.calls.length).toBe(1);
-            expect(longItem.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("returns no match when multiple items have identical shortest exact labels", () => {
-            let item1 = {label: "TV", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let item2 = {label: "TV", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            openhab.items.getItems.mockReturnValue([item1, item2]);
-
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            // Both items have the exact same shortest label -> ambiguous -> no command sent
-            rbi.interpretUtterance("TV foo");
-            expect(item1.sendCommand.mock.calls.length).toBe(0);
-            expect(item2.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("chooses the item with the shortest exact synonym when labels are longer", () => {
-            let getMetadataImplementation = (data) => (namespace) => {
-                if (namespace != "synonyms") {
-                    return null;
-                }
-                return {value: data};
-            };
-
-            let itemShortSyn = {label: "Living Room TV", sendCommand: jest.fn(), getMetadata: jest.fn(getMetadataImplementation("TV"))};
-            let itemOther = {label: "TV Room", sendCommand: jest.fn(), getMetadata: jest.fn(getMetadataImplementation("TV Room"))};
-            openhab.items.getItems.mockReturnValue([itemShortSyn, itemOther]);
-
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(), opt("Room"), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            // Utterance "TV Room foo" should match both (synonym "TV" and synonym "TV Room"),
-            // but the shortest exact synonym ("TV") wins -> itemShortSyn should be called.
-            rbi.interpretUtterance("TV Room foo");
-            expect(itemShortSyn.sendCommand.mock.calls.length).toBe(1);
-            expect(itemOther.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("is ambiguous when a label and another item's synonym are equally short exact matches", () => {
-            // item1 has label "TV" (short), item2 has synonym "TV" (short via synonym)
-            let item1 = {label: "TV", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let getMetadataImplementation = (data) => (namespace) => {
-                if (namespace != "synonyms") {
-                    return null;
-                }
-                return {value: data};
-            };
-            let item2 = {label: "Television", sendCommand: jest.fn(), getMetadata: jest.fn(getMetadataImplementation("TV"))};
-            openhab.items.getItems.mockReturnValue([item1, item2]);
-
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            // Both provide a shortest exact match of length 1 -> ambiguous -> no command sent
-            rbi.interpretUtterance("TV foo");
-            expect(item1.sendCommand.mock.calls.length).toBe(0);
-            expect(item2.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("handles items missing getMetadata method and still finds shortest match", () => {
-            // item1 has no getMetadata; item2 has a synonym "TV Set" -> item1 should be selected for "TV"
-            let item1 = {label: "TV", sendCommand: jest.fn()}; // No getMetadata
-            let getMetadataImplementation = (data) => (namespace) => {
-                if (namespace != "synonyms") {
-                    return null;
-                }
-                return {value: data};
-            };
-            let item2 = {label: "Television", sendCommand: jest.fn(), getMetadata: jest.fn(getMetadataImplementation("TV Set"))};
-            openhab.items.getItems.mockReturnValue([item1, item2]);
-
-            let cmdParameter = 123;
-            let testExpression = seq(itemLabel(), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            rbi.interpretUtterance("TV foo");
-            expect(item1.sendCommand.mock.calls.length).toBe(1);
-            expect(item2.sendCommand.mock.calls.length).toBe(0);
-        });
-    });
-
-    describe("optional expression", () => {
-        it("matches same or none string", () => {
-            let testExpression = seq(opt("bar"), "foo");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("foo");
-            expect(testFunction.mock.calls.length).toBe(1);
-            rbi.interpretUtterance("bar foo");
-            expect(testFunction.mock.calls.length).toBe(2);
-        });
-
-        it("does not match other string", () => {
-            let testExpression = seq(opt("bar"), "foo");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("other foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            rbi.interpretUtterance("bar bar foo");
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-
-        it("is successful at the end when no tokens are left", () => {
-            let testExpression = seq("first", alt("second", opt("foo")), opt("bar"));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("first second");
-            expect(testFunction.mock.calls.length).toBe(1);
-        });
-    });
-
-    describe("single sequence expression", () => {
-        it("matches correct sequence", () => {
-            let testExpression = seq("bar", "foo", "foobar");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("bar foo foobar");
-            expect(testFunction.mock.calls.length).toBe(1);
-        });
-
-        it("does not match wrong sequence", () => {
-            let testExpression = seq("bar", "foo", "foobar");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("foo foobar");
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-
-        it("matches normalized unicode characters and apostrophes", () => {
-            let testExpression = seq("tÈS$Tö'ó");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("2TèstÖ'Ó");
-            expect(testFunction.mock.calls.length).toBe(1);
-        });
-
-        it("matches first token and ignores further", () => {
-            let testExpression = seq("foo bar");
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("foo abc");
-            expect(testFunction.mock.calls.length).toBe(1);
-        });
-    });
-
-    describe("single string expression", () => {
-        it("matches same string and executes rule function", () => {
-            let testExpression = "foo";
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("foo");
-            expect(testFunction.mock.calls.length).toBe(1);
-        });
-
-        it("does not match similar string", () => {
-            let testExpression = "foo";
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("fo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            rbi.interpretUtterance("fooo");
-            expect(testFunction.mock.calls.length).toBe(0);
-            rbi.interpretUtterance("fo'o");
-            expect(testFunction.mock.calls.length).toBe(0);
-            rbi.interpretUtterance("foo'");
-            expect(testFunction.mock.calls.length).toBe(0);
-            rbi.interpretUtterance("fooö");
-            expect(testFunction.mock.calls.length).toBe(0);
-            rbi.interpretUtterance("fooò");
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-        
-        describe("annotation result", () => {
-            it("calls function after match but returns unsuccessful if function fails", () => {
-                let testExpression = seq("bar", "foo", "foobar");
-                let testFunction = jest.fn();
-                testFunction.mockReturnValue(false);
-                rbi.addRule(testExpression, testFunction);
-                var result = rbi.interpretUtterance("bar foo foobar");
-                expect(result.success).toBe(false);
-                expect(testFunction.mock.calls.length).toBe(1);
-            });
-
-            it("annotates named function", () => {
-                let testExpression = seq("bar", "foo", "foobar");
-                let testFunction = function(){return true;};
-                rbi.addRule(testExpression, testFunction);
-                var result = rbi.interpretUtterance("bar foo foobar");
-                expect(result.success).toBe(true);
-                expect(result.executeFunction).toBe(testFunction.name);
-            });
-
-            it("annotates anonymous function", () => {
-                let testExpression = seq("bar", "foo", "foobar");
-                rbi.addRule(testExpression, function(myMarkerParam){return true;});
-                var result = rbi.interpretUtterance("bar foo foobar");
-                expect(result.success).toBe(true);
-                expect(result.executeFunction).toContain("return true;");
-                expect(result.executeFunction).toContain("myMarkerParam");
-            });
-        });
-
-        it("does not match other string", () => {
-            let testExpression = "foo";
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("bar");
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-        
-        it("does not match empty string", () => {
-            let testExpression = "foo";
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance("");
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-        
-        it("does not match null", () => {
-            let testExpression = "foo";
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance(null);
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-        
-        it("does not match undefined", () => {
-            let testExpression = "foo";
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-            rbi.interpretUtterance(undefined);
-            expect(testFunction.mock.calls.length).toBe(0);
-        });
-    });
-
-    describe("itemLabel isGroup behavior", () => {
-        it("matches group when isGroup === true even if non-group shares label", () => {
-            // Both items share the same label; when isGroup === true the group should be selected.
-            let sharedLabel = "shared label";
-            let groupItem = {label: sharedLabel, type: "Group", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let normalItem = {label: sharedLabel, type: "Switch", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            openhab.items.getItems.mockReturnValue([groupItem, normalItem]);
-
-            let cmdParameter = 321;
-            let testExpression = seq(itemLabel(true, true), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            rbi.interpretUtterance(`${sharedLabel} foo`);
-            // Only the group should be considered due to isGroup === true
-            expect(groupItem.sendCommand.mock.calls.length).toBe(1);
-            expect(groupItem.sendCommand.mock.calls[0][0]).toBe(cmdParameter);
-            expect(normalItem.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("is ambiguous (no match) when isGroup === false and duplicate labels exist", () => {
-            // When isGroup === false both normal and group items are considered -> ambiguity -> no match
-            let sharedLabel = "shared label";
-            let groupItem = {label: sharedLabel, type: "Group", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let normalItem = {label: sharedLabel, type: "Switch", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            openhab.items.getItems.mockReturnValue([groupItem, normalItem]);
-
-            let cmdParameter = 456;
-            let testExpression = seq(itemLabel(true, false), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            rbi.interpretUtterance(`${sharedLabel} foo`);
-            // No single exact match should be found -> no sendCommand calls
-            expect(normalItem.sendCommand.mock.calls.length).toBe(0);
-            expect(groupItem.sendCommand.mock.calls.length).toBe(0);
-        });
-
-        it("is ambiguous (no match) when isGroup is omitted (null) and duplicate labels exist", () => {
-            let sharedLabel = "shared label";
-            let groupItem = {label: sharedLabel, type: "Group", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            let normalItem = {label: sharedLabel, type: "Switch", sendCommand: jest.fn(), getMetadata: jest.fn()};
-            openhab.items.getItems.mockReturnValue([groupItem, normalItem]);
-
-            let cmdParameter = 789;
-            let testExpression = seq(itemLabel(), cmd("foo", cmdParameter));
-            let testFunction = jest.fn();
-            rbi.addRule(testExpression, testFunction);
-
-            rbi.interpretUtterance(`${sharedLabel} foo`);
-            // With omitted isGroup (null) both items are considered -> ambiguity -> no commands
-            expect(normalItem.sendCommand.mock.calls.length).toBe(0);
-            expect(groupItem.sendCommand.mock.calls.length).toBe(0);
-        });
-    });
-});
